requirements.txt
openai>=1.6.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=4.5.0
httpx>=0.25.0
pydantic-settings>=2.0.0
psutil>=5.9.0
structlog>=23.0.0
python-multipart>=0.0.6
plotly>=5.0.0 
Jinja2>=3.1.0
google-cloud-pubsub>=2.31.1
google-cloud-storage>=3.3.1
google-cloud-bigquery>=3.37.0
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0
google-api-python-client>=2.181.0
google-cloud-resource-manager>=1.14.2